requirements.txt
discord.py[voice]
discord.py
urllib3
yt-dlp
python-dotenv
mistralai 

#PyNaCl